Feat(docker): Dockerizacion del proyecto

+++ pis-ecocontrol-6/hooks/SessionUtil.js
@@ -1,24 +1,41 @@
+"use client";
+
 export const save = (key, data) => {
-    sessionStorage.setItem(key, data);
+    if (typeof window !== 'undefined') {
+        sessionStorage.setItem(key, data);
+    }
 }
 
 export const get = (key) => {
-   return sessionStorage.getItem(key, data);
+    if (typeof window !== 'undefined') {
+        return sessionStorage.getItem(key);
+    }
+    return null;
 }
 
 export const saveToken = (key) => {
-    sessionStorage.setItem("token",key);
+    if (typeof window !== 'undefined') {
+        sessionStorage.setItem("token", key);
+    }
 }
 
 export const getToken = () => {
-    return sessionStorage.getItem("token");
+    if (typeof window !== 'undefined') {
+        return sessionStorage.getItem("token");
+    }
+    return null;
 }
 
 export const borrarSesion = () => {
-    sessionStorage.clear();
+    if (typeof window !== 'undefined') {
+        sessionStorage.clear();
+    }
 }
 
 export const estaSesion = () => {
-    var token = sessionStorage.getItem('token');
-    return (token && (token != 'undefined' || token != null || token != 'null'));
+    if (typeof window !== 'undefined') {
+        var token = sessionStorage.getItem('token');
+        return (token && (token !== 'undefined' && token !== null && token !== 'null'));
+    }
+    return false;
 }
